text-minimessage: builtin -> standard
