+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/MiniMessageSerializer.java
@@ -38,14 +38,14 @@ import net.kyori.adventure.text.format.Style;
 import net.kyori.adventure.text.format.TextColor;
 import net.kyori.adventure.text.format.TextDecoration;
 import net.kyori.adventure.text.minimessage.parser.TokenParser;
-import net.kyori.adventure.text.minimessage.tag.builtin.ClickTag;
-import net.kyori.adventure.text.minimessage.tag.builtin.ColorTagResolver;
-import net.kyori.adventure.text.minimessage.tag.builtin.DecorationTag;
-import net.kyori.adventure.text.minimessage.tag.builtin.FontTag;
-import net.kyori.adventure.text.minimessage.tag.builtin.HoverTag;
-import net.kyori.adventure.text.minimessage.tag.builtin.InsertionTag;
-import net.kyori.adventure.text.minimessage.tag.builtin.KeybindTag;
-import net.kyori.adventure.text.minimessage.tag.builtin.TranslatableTag;
+import net.kyori.adventure.text.minimessage.tag.standard.ClickTag;
+import net.kyori.adventure.text.minimessage.tag.standard.ColorTagResolver;
+import net.kyori.adventure.text.minimessage.tag.standard.DecorationTag;
+import net.kyori.adventure.text.minimessage.tag.standard.FontTag;
+import net.kyori.adventure.text.minimessage.tag.standard.HoverTag;
+import net.kyori.adventure.text.minimessage.tag.standard.InsertionTag;
+import net.kyori.adventure.text.minimessage.tag.standard.KeybindTag;
+import net.kyori.adventure.text.minimessage.tag.standard.TranslatableTag;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
