+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/tag/standard/DecorationTag.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.minimessage.tag.builtin;
+package net.kyori.adventure.text.minimessage.tag.standard;
 
 import java.util.Collections;
 import java.util.HashMap;
