+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/tag/standard/RainbowTag.java
@@ -21,20 +21,15 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.minimessage.tag.builtin;
+package net.kyori.adventure.text.minimessage.tag.standard;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 import java.util.Objects;
-import java.util.OptionalDouble;
 import java.util.PrimitiveIterator;
 import java.util.stream.Stream;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TextComponent;
 import net.kyori.adventure.text.flattener.ComponentFlattener;
-import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextColor;
 import net.kyori.adventure.text.minimessage.Context;
 import net.kyori.adventure.text.minimessage.parser.node.TagNode;
@@ -50,81 +45,51 @@ import net.kyori.examination.ExaminableProperty;
 import org.jetbrains.annotations.NotNull;
 
 /**
- * A transformation that applies a colour gradient.
+ * Applies rainbow color to a component.
  *
  * @since 4.10.0
  */
-public final class GradientTag implements Modifying, Examinable {
-  public static final String GRADIENT = "gradient";
+public final class RainbowTag implements Modifying, Examinable {
+  private static final String REVERSE = "!";
+  public static final String RAINBOW = "rainbow";
 
-  private int size = 0;
+  private int size;
   private int disableApplyingColorDepth = -1;
 
-  private int index = 0;
   private int colorIndex = 0;
 
-  private float factorStep = 0;
-  private final TextColor[] colors;
-  private float phase;
-  private final boolean negativePhase;
+  private float center = 128;
+  private float width = 127;
+  private double frequency = 1;
+  private final boolean reversed;
+
+  private final int phase;
 
   static Tag create(final ArgumentQueue args, final Context ctx) {
-    float phase = 0;
-    final List<TextColor> textColors;
+    boolean reversed = false;
+    int phase = 0;
+
     if (args.hasNext()) {
-      textColors = new ArrayList<>();
-      while (args.hasNext()) {
-        final Tag.Argument arg = args.pop();
-        // last argument? maybe this is the phase?
-        if (!args.hasNext()) {
-          final OptionalDouble possiblePhase = arg.asDouble();
-          if (possiblePhase.isPresent()) {
-            phase = (float) possiblePhase.getAsDouble();
-            if (phase < -1f || phase > 1f) {
-              throw ctx.newError(String.format("Gradient phase is out of range (%s). Must be in the range [-1.0f, 1.0f] (inclusive).", phase), args);
-            }
-            break;
-          }
-        }
-
-        final String argValue = arg.value();
-        final TextColor parsedColor;
-        if (argValue.charAt(0) == '#') {
-          parsedColor = TextColor.fromHexString(argValue);
-        } else {
-          parsedColor = NamedTextColor.NAMES.value(arg.lowerValue());
-        }
-        if (parsedColor == null) {
-          throw ctx.newError(String.format("Unable to parse a color from '%s'. Please use named colours or hex (#RRGGBB) colors.", argValue), args);
-        }
-        textColors.add(parsedColor);
+      String value = args.pop().value();
+      if (value.startsWith(REVERSE)) {
+        reversed = true;
+        value = value.substring(REVERSE.length());
       }
-
-      if (textColors.size() < 2) {
-        throw ctx.newError("Invalid gradient, not enough colors. Gradients must have at least two colors.", args);
+      if (value.length() > 0) {
+        try {
+          phase = Integer.parseInt(value);
+        } catch (final NumberFormatException ex) {
+          throw ctx.newError("Expected phase, got " + value);
+        }
       }
-    } else {
-      textColors = Collections.emptyList();
     }
 
-    return new GradientTag(phase, textColors);
+    return new RainbowTag(reversed, phase);
   }
 
-  private GradientTag(final float phase, final List<TextColor> colors) {
-    if (phase < 0) {
-      this.negativePhase = true;
-      this.phase = 1 + phase;
-      Collections.reverse(colors);
-    } else {
-      this.negativePhase = false;
-      this.phase = phase;
-    }
-
-    if (colors.isEmpty()) {
-      this.colors = new TextColor[]{TextColor.color(0xffffff), TextColor.color(0x000000)};
-    } else {
-      this.colors = colors.toArray(new TextColor[0]);
-    }
+  private RainbowTag(final boolean reversed, final int phase) {
+    this.reversed = reversed;
+    this.phase = phase;
   }
 
   @Override
@@ -135,7 +100,7 @@ public final class GradientTag implements Modifying, Examinable {
     } else if (curr instanceof TagNode) {
       final TagNode tag = (TagNode) curr;
       if (tag.tag() instanceof Inserting) {
-        // ComponentTransformation.apply() returns the value of the component placeholder
+        // Inserting.apply() returns the value of the component placeholder
         ComponentFlattener.textOnly().flatten(((Inserting) tag.tag()).value(), s -> this.size += s.codePointCount(0, s.length()));
       }
     }
@@ -144,13 +109,9 @@ public final class GradientTag implements Modifying, Examinable {
   @Override
   public void postVisit() {
     // init
-    int sectorLength = this.size / (this.colors.length - 1);
-    if (sectorLength < 1) {
-      sectorLength = 1;
-    }
-    this.factorStep = 1.0f / (sectorLength + this.index);
-    this.phase = this.phase * sectorLength;
-    this.index = 0;
+    this.center = 128;
+    this.width = 127;
+    this.frequency = Math.PI * 2 / this.size;
   }
 
   @Override
@@ -166,7 +127,7 @@ public final class GradientTag implements Modifying, Examinable {
         final int len = content.codePointCount(0, content.length());
         for (int i = 0; i < len; i++) {
           // increment our color index
-          this.color();
+          this.color(this.phase);
         }
       }
       return current.children(Collections.emptyList());
@@ -179,64 +140,51 @@ public final class GradientTag implements Modifying, Examinable {
 
       final TextComponent.Builder parent = Component.text();
 
+      if (this.colorIndex == 0 && this.reversed) {
+        this.colorIndex = this.size - 1;
+      }
+
       // apply
       final int[] holder = new int[1];
       for (final PrimitiveIterator.OfInt it = content.codePoints().iterator(); it.hasNext();) {
         holder[0] = it.nextInt();
-        final Component comp = Component.text(new String(holder, 0, 1), this.color());
+        final Component comp = Component.text(new String(holder, 0, 1), this.color(this.phase));
         parent.append(comp);
       }
 
       return parent.build();
     }
 
-    return Component.empty().mergeStyle(current);
+    return Component.text("", current.style());
   }
 
-  private TextColor color() {
-    // color switch needed?
-    if (this.factorStep * this.index > 1) {
-      this.colorIndex++;
-      this.index = 0;
-    }
-
-    float factor = this.factorStep * (this.index++ + this.phase);
-    // loop around if needed
-    if (factor > 1) {
-      factor = 1 - (factor - 1);
-    }
-
-    if (this.negativePhase && this.colors.length % 2 != 0) {
-      // flip the gradient segment for to allow for looping phase -1 through 1
-      return TextColor.lerp(factor, this.colors[this.colorIndex + 1], this.colors[this.colorIndex]);
-    } else {
-      return TextColor.lerp(factor, this.colors[this.colorIndex], this.colors[this.colorIndex + 1]);
-    }
+  private TextColor color(final float phase) {
+    final int index = this.reversed ? this.colorIndex-- : this.colorIndex++;
+    final int red = (int) (Math.sin(this.frequency * index + 2 + phase) * this.width + this.center);
+    final int green = (int) (Math.sin(this.frequency * index + 0 + phase) * this.width + this.center);
+    final int blue = (int) (Math.sin(this.frequency * index + 4 + phase) * this.width + this.center);
+    return TextColor.color(red, green, blue);
   }
 
   @Override
   public @NotNull Stream<? extends ExaminableProperty> examinableProperties() {
-    return Stream.of(
-      ExaminableProperty.of("phase", this.phase),
-      ExaminableProperty.of("colors", this.colors)
-    );
+    return Stream.of(ExaminableProperty.of("phase", this.phase));
   }
 
   @Override
   public boolean equals(final Object other) {
     if (this == other) return true;
     if (other == null || this.getClass() != other.getClass()) return false;
-    final GradientTag that = (GradientTag) other;
-    return this.index == that.index
-      && this.colorIndex == that.colorIndex
-      && ShadyPines.equals(that.factorStep, this.factorStep)
-      && this.phase == that.phase && Arrays.equals(this.colors, that.colors);
+    final RainbowTag that = (RainbowTag) other;
+    return this.colorIndex == that.colorIndex
+      && ShadyPines.equals(that.center, this.center)
+      && ShadyPines.equals(that.width, this.width)
+      && ShadyPines.equals(that.frequency, this.frequency)
+      && this.phase == that.phase;
   }
 
   @Override
   public int hashCode() {
-    int result = Objects.hash(this.index, this.colorIndex, this.factorStep, this.phase);
-    result = 31 * result + Arrays.hashCode(this.colors);
-    return result;
+    return Objects.hash(this.colorIndex, this.center, this.width, this.frequency, this.phase);
   }
 }
